Factor out SelectableBtn
